+++ src/crawlee/storage_clients/models.py
@@ -137,7 +137,9 @@ class ProcessedRequest(BaseModel):
 
     model_config = ConfigDict(populate_by_name=True)
 
-    id: Annotated[str, Field(alias='requestId')]
+    id: Annotated[str | None, Field(alias='requestId', default=None)] = None
+    """Internal representation of the request by the storage client. Only some clients use id."""
+
     unique_key: Annotated[str, Field(alias='uniqueKey')]
     was_already_present: Annotated[bool, Field(alias='wasAlreadyPresent')]
     was_already_handled: Annotated[bool, Field(alias='wasAlreadyHandled')]